Split tests into server, client and common
